+++ src/main/java/constants/game/GameConstants.java
@@ -1,26 +1,18 @@
 package constants.game;
 
 import client.MapleDisease;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.HashMap;
-import java.util.Map;
 import client.MapleJob;
 import config.YamlConfig;
 import constants.skills.Aran;
+import provider.*;
+import server.maps.FieldLimit;
+import server.maps.MapleMap;
+import server.quest.MapleQuest;
+
 import java.io.File;
 import java.text.DecimalFormat;
 import java.text.NumberFormat;
-import java.util.Locale;
-import provider.MapleData;
-import provider.MapleDataDirectoryEntry;
-import provider.MapleDataFileEntry;
-import provider.MapleDataProvider;
-import provider.MapleDataProviderFactory;
-import provider.MapleDataTool;
-import server.maps.MapleMap;
-import server.maps.FieldLimit;
-import server.quest.MapleQuest;
+import java.util.*;
 
 /*
  * @author kevintjuh93
@@ -518,8 +510,8 @@ public class GameConstants {
     private static boolean isInBranchJobTree(int skillJobId, int jobId, int branchType) {
         int branch = (int)(Math.pow(10, branchType));
         
-        int skillBranch = (int)(skillJobId / branch) * branch;
-        int jobBranch = (int)(jobId / branch) * branch;
+        int skillBranch = (skillJobId / branch) * branch;
+        int jobBranch = (jobId / branch) * branch;
         
         return skillBranch == jobBranch;
     }
@@ -527,8 +519,8 @@ public class GameConstants {
     private static boolean hasDivergedBranchJobTree(int skillJobId, int jobId, int branchType) {
         int branch = (int)(Math.pow(10, branchType));
         
-        int skillBranch = (int)(skillJobId / branch);
-        int jobBranch = (int)(jobId / branch);
+        int skillBranch = skillJobId / branch;
+        int jobBranch = jobId / branch;
         
         return skillBranch != jobBranch && skillBranch % 10 != 0;
     }
